fix: positioning logic
use getBoundingClientRect instead of offset properites, this handles transforms and SVG elements
