+++ src/dom-utils/getBoundingClientRect.js
@@ -1,6 +1,6 @@
 // @flow
 import type { ClientRectObject } from '../types';
 
-export default (element: HTMLElement): ClientRectObject => {
+export default (element: Element): ClientRectObject => {
   return JSON.parse(JSON.stringify(element.getBoundingClientRect()));
 };
