+++ src/dom-utils/getElementClientRect.js
@@ -1,13 +1,19 @@
 // @flow
 import type { Rect } from '../types';
+import getBoundingClientRect from './getBoundingClientRect';
+import listScrollParents from './listScrollParents';
+import getScrollSum from './getScrollSum';
 
-// Returns the width, height and offsets of the provided element
+// Returns the width, height and offsets of the provided element relative to the
+// offsetParent
 export default (element: HTMLElement): Rect => {
-  // get the basic client rect, it doesn't include margins
-  const width = element.offsetWidth;
-  const height = element.offsetHeight;
-  const y = element.offsetTop;
-  const x = element.offsetLeft;
+  const rect = getBoundingClientRect(element);
+  const { scrollLeft, scrollTop } = getScrollSum(listScrollParents(element));
 
-  return { width, height, y, x };
+  const width = rect.width;
+  const height = rect.height;
+  const x = rect.left + scrollLeft;
+  const y = rect.top + scrollTop;
+
+  return { width, height, x, y };
 };
