+++ src/dom-utils/getScrollSum.js
@@ -1,0 +1,13 @@
+// @flow
+import getNodeScroll from './getNodeScroll';
+
+export default (scrollParents: Array<Node>) =>
+  scrollParents.reduce(
+    (scroll, scrollParent) => {
+      const nodeScroll = getNodeScroll(scrollParent);
+      scroll.scrollTop += nodeScroll.scrollTop;
+      scroll.scrollLeft += nodeScroll.scrollLeft;
+      return scroll;
+    },
+    { scrollTop: 0, scrollLeft: 0 }
+  );
